Use Maven build system to match JMH

+++ src/main/java/cz/cuni/mff/d3s/spl/data/BenchmarkRunUtils.java
@@ -37,7 +37,7 @@ public class BenchmarkRunUtils  {
 	}
 	
 	public static interface Transformer {
-		public long apply(long sample);
+		public Double apply(Double sample);
 	}
 	
 	/** Compute mean of a run. */
@@ -137,7 +137,7 @@ public class BenchmarkRunUtils  {
 	
 	public static BenchmarkRun transform(final BenchmarkRun run, Transformer transformer) {
 		BenchmarkRunBuilder builder = new BenchmarkRunBuilder();
-		for (long sample : run.getSamples()) {
+		for (Double sample : run.getSamples()) {
 			builder.addSamples(transformer.apply(sample));
 		}
 		return builder.create();
