+++ src/test/java/cz/cuni/mff/d3s/spl/data/BenchmarkRunUtilsTest.java
@@ -80,7 +80,7 @@ public class BenchmarkRunUtilsTest {
 		}
 
 		@Override
-		public long apply(long sample) {
+		public Double apply(Double sample) {
 			return sample * mult;
 		}
 	}
